infrastructure/webdriver/bidi/no-bidi.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test testdriver BiDi is not available if not included</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
    promise_test(async () => {
        assert_equals(test_driver.bidi, undefined);
    }, "Assert testdriver BiDi API is not available");
</script>
